+++ CHANGELOG.md
@@ -1,17 +1,18 @@
 <!-- usage documentation: http://expeditor-docs.es.chef.io/configuration/changelog/ -->
 This changelog lists individual merged pull requests to Chef Infra Client and geared towards developers. For a list of significant changes per release see the [Chef Infra Client Release Notes](https://docs.chef.io/release_notes_client/).
 
-<!-- latest_release 19.4.9 -->
-## [v19.4.9](https://github.com/chef/chef/tree/v19.4.9) (2026-07-30)
+<!-- latest_release 19.4.10 -->
+## [v19.4.10](https://github.com/chef/chef/tree/v19.4.10) (2026-07-31)
 
 #### Merged Pull Requests
-- Temporarily skip habitat_install archive_file in Windows kitchen [#16249](https://github.com/chef/chef/pull/16249) ([tpowell-progress](https://github.com/tpowell-progress))
+- use rmdir instead of remove-item for windows plans [#16238](https://github.com/chef/chef/pull/16238) ([rishichawda](https://github.com/rishichawda))
 <!-- latest_release -->
 
 <!-- release_rollup since=19.3.15 -->
 ### Changes not yet released to stable
 
 #### Merged Pull Requests
+- use rmdir instead of remove-item for windows plans [#16238](https://github.com/chef/chef/pull/16238) ([rishichawda](https://github.com/rishichawda)) <!-- 19.4.10 -->
 - Temporarily skip habitat_install archive_file in Windows kitchen [#16249](https://github.com/chef/chef/pull/16249) ([tpowell-progress](https://github.com/tpowell-progress)) <!-- 19.4.9 -->
 - Widen ffi-yajl version constraint [#16240](https://github.com/chef/chef/pull/16240) ([tpowell-progress](https://github.com/tpowell-progress)) <!-- 19.4.8 -->
 - Temporarily skip archive_file test in Windows kitchen [#16247](https://github.com/chef/chef/pull/16247) ([neha-p6](https://github.com/neha-p6)) <!-- 19.4.7 -->